+++ CHANGES.md
@@ -19,12 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   (an stp file contains the starting position of the corresponding proof)
 - command theorem to generate the lp files corresponding to a named theorem
 - Makefile to generate and check lp and coq files generated with split
-- fusion.ml: do not generate new theorems for empty instantiations
+- command size to get statistics on the size of terms
+- option --sharing for using maximal sharing when recording term abbreviations
+- option --print-stats to print statistics on hash tables at exit
 
 ### Modified
 
 - identifier renamings
 - merged the command dg in the command mk
+- fusion.ml: do not generate new theorems for empty instantiations
 
 ## 0.0.1 (2023-11-22)